requirements.txt
Django
djangorestframework
django-filter
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
environs
